feat: implement injective demo
