+++ packages/dapp-kit/src/contexts/index.ts
@@ -1,9 +1,10 @@
 import { createContext } from 'react';
-import type { Sodax } from '@sodax/sdk';
+import type { EvmHubProvider, Sodax } from '@sodax/sdk';
 
 export interface SodaxContextType {
   sodax: Sodax;
   testnet: boolean;
+  hubProvider: EvmHubProvider | undefined;
 }
 
 export const SodaxContext = createContext<SodaxContextType | null>(null);
