+++ packages/dapp-kit/src/providers/SodaxProvider.tsx
@@ -1,7 +1,7 @@
 import type { ReactNode, ReactElement } from 'react';
-import { Sodax, type SodaxConfig } from '@sodax/sdk';
+import { EvmHubProvider, getHubChainConfig, Sodax, type SodaxConfig } from '@sodax/sdk';
 import { SodaxContext } from '@/contexts';
-import React from 'react';
+import React, { useMemo } from 'react';
 
 interface SodaxProviderProps {
   children: ReactNode;
@@ -12,5 +12,20 @@ interface SodaxProviderProps {
 export const SodaxProvider = ({ children, testnet = false, config }: SodaxProviderProps): ReactElement => {
   const sodax = new Sodax(config);
 
-  return <SodaxContext.Provider value={{ sodax, testnet }}>{children}</SodaxContext.Provider>;
+  const hubChainId = config?.hubProviderConfig?.chainConfig.chain.id;
+  const hubRpcUrl = config?.hubProviderConfig?.hubRpcUrl;
+
+  const hubProvider = useMemo(() => {
+    if (hubChainId && hubRpcUrl) {
+      const hubChainCfg = getHubChainConfig(hubChainId);
+
+      return new EvmHubProvider({
+        hubRpcUrl: hubRpcUrl,
+        chainConfig: hubChainCfg,
+      });
+    }
+    return undefined;
+  }, [hubChainId, hubRpcUrl]);
+
+  return <SodaxContext.Provider value={{ sodax, testnet, hubProvider }}>{children}</SodaxContext.Provider>;
 };
